+++ crates/lib/src/qpu/translation.rs
@@ -8,18 +8,24 @@ use qcs_api_client_grpc::{
     services::translation::{
         translate_quil_to_encrypted_controller_job_request::NumShots,
         translation_options::TranslationBackend, BackendV1Options, BackendV2Options,
+        GetQuantumProcessorQuilCalibrationProgramRequest,
         TranslateQuilToEncryptedControllerJobRequest, TranslationOptions as ApiTranslationOptions,
     },
 };
-use qcs_api_client_openapi::{
-    apis::{translation_api, Error as OpenAPIError},
-    models::GetQuiltCalibrationsResponse,
-};
 use tokio::time::error::Elapsed;
-#[cfg(feature = "tracing")]
-use tracing::instrument;
 
 use crate::client::{GrpcClientError, Qcs, DEFAULT_HTTP_API_TIMEOUT};
+
+/// Errors that can occur when making a request to translation service.
+#[derive(Debug, thiserror::Error)]
+pub enum Error {
+    /// Error due to gRPC client
+    #[error(transparent)]
+    Grpc(#[from] GrpcClientError),
+    /// Error due to client timeout
+    #[error("Client configured timeout exceeded")]
+    ClientTimeout(#[from] Elapsed),
+}
 
 /// An encrypted and translated program, along with `readout_map`
 /// to map job `readout_data` back to program-declared variables.
@@ -36,14 +42,13 @@ pub struct EncryptedTranslationResult {
 }
 
 /// Translate a program, returning an encrypted and translated program.
-#[cfg_attr(feature = "tracing", instrument(skip_all))]
 pub async fn translate<TO>(
     quantum_processor_id: &str,
     quil_program: &str,
     num_shots: u32,
     client: &Qcs,
     translation_options: Option<TO>,
-) -> Result<EncryptedTranslationResult, GrpcClientError>
+) -> Result<EncryptedTranslationResult, Error>
 where
     TO: Into<ApiTranslationOptions>,
 {
@@ -64,9 +69,11 @@ where
     };
 
     let response = client
-        .get_translation_client()?
+        .get_translation_client()
+        .map_err(GrpcClientError::from)?
         .translate_quil_to_encrypted_controller_job(request)
-        .await?
+        .await
+        .map_err(GrpcClientError::from)?
         .into_inner();
 
     Ok(EncryptedTranslationResult {
@@ -80,36 +87,33 @@ where
     })
 }
 
-/// API Errors encountered when trying to get Quil-T calibrations.
-#[derive(Debug, thiserror::Error)]
-pub enum GetQuiltCalibrationsError {
-    /// Failed the http call
-    #[error("Failed to get Quil-T calibrations via API: {0}")]
-    ApiError(#[from] OpenAPIError<translation_api::GetQuiltCalibrationsError>),
-
-    /// API call did not finish before timeout
-    #[error("API call did not finish before timeout.")]
-    TimeoutError(#[from] Elapsed),
-}
-
 /// Query the QCS API for Quil-T calibrations.
 /// If `None`, the default `timeout` used is 10 seconds.
 pub async fn get_quilt_calibrations(
-    quantum_processor_id: &str,
+    quantum_processor_id: String,
     client: &Qcs,
     timeout: Option<Duration>,
-) -> Result<GetQuiltCalibrationsResponse, GetQuiltCalibrationsError> {
+) -> Result<String, Error> {
     #[cfg(feature = "tracing")]
     tracing::debug!("getting Quil-T calibrations for {}", quantum_processor_id);
 
     let timeout = timeout.unwrap_or(DEFAULT_HTTP_API_TIMEOUT);
 
+    let mut translation_client = client
+        .get_translation_client()
+        .map_err(GrpcClientError::from)?;
+
     tokio::time::timeout(timeout, async move {
-        Ok(translation_api::get_quilt_calibrations(
-            &client.get_openapi_client(),
-            quantum_processor_id,
-        )
-        .await?)
+        Ok(translation_client
+            .get_quantum_processor_quil_calibration_program(
+                GetQuantumProcessorQuilCalibrationProgramRequest {
+                    quantum_processor_id,
+                },
+            )
+            .await
+            .map_err(GrpcClientError::from)?
+            .into_inner()
+            .quil_calibration_program)
     })
     .await?
 }
@@ -177,6 +181,7 @@ impl TranslationOptions {
         options
     }
 
+    #[allow(dead_code)]
     fn ensure_backend_v1(&mut self) -> Result<&mut BackendV1Options, TranslationBackendMismatch> {
         if matches!(self.backend(), None | Some(TranslationBackend::V1(_))) {
             Ok(self.with_backend_v1())